frontend/explorer: decouple explorer and flyout browsing paths from current_path
Explorer and flyout now maintain independent browsing paths
(explorer_browsing_path, flyout_browsing_path) in the Redux store,
separate from the project-wide current_path which tracks the active
file context.  A "Switch" button appears when the browsing path
diverges from current_path.

Also extracts ~750 lines from explorer.tsx into directory-tree.tsx,
~170 lines from flyouts/files.tsx into use-computed-files.ts, and
deduplicates flyout navigation into use-flyout-navigation.ts.
