+++ src/packages/frontend/project/explorer/path-navigator.tsx
@@ -21,6 +21,22 @@ interface Props {
   style?: React.CSSProperties;
   className?: string;
   mode?: "files" | "flyout";
+  /**
+   * Override the browsing path shown in the breadcrumb.
+   * When omitted, reads `current_path` from the Redux store.
+   */
+  currentPath?: string;
+  /**
+   * Override the history path used for breadcrumb depth.
+   * When omitted, reads `history_path` from the Redux store.
+   */
+  historyPath?: string;
+  /**
+   * Called instead of `actions.open_directory` when the user clicks a
+   * breadcrumb segment or the up/home buttons.  When omitted, falls
+   * back to `actions.open_directory(path, true, false)`.
+   */
+  onNavigate?: (path: string) => void;
 }
 
 // This path consists of several PathSegmentLinks
@@ -31,10 +47,22 @@ export const PathNavigator: React.FC<Props> = React.memo(
       style,
       className = "cc-path-navigator",
       mode = "files",
+      onNavigate,
     } = props;
-    const currentPath = useTypedRedux({ project_id }, "current_path");
-    const historyPath = useTypedRedux({ project_id }, "history_path");
+    const reduxCurrentPath = useTypedRedux({ project_id }, "current_path");
+    const reduxHistoryPath = useTypedRedux({ project_id }, "history_path");
     const actions = useActions({ project_id });
+
+    const currentPath = props.currentPath ?? reduxCurrentPath;
+    const historyPath = props.historyPath ?? reduxHistoryPath;
+
+    const navigate = (path: string) => {
+      if (onNavigate) {
+        onNavigate(path);
+      } else {
+        actions?.open_directory(path, true, false);
+      }
+    };
 
     function make_path() {
       const v: any[] = [];
@@ -67,7 +95,7 @@ export const PathNavigator: React.FC<Props> = React.memo(
           ),
           full_name: "",
           key: 0,
-          on_click: () => actions?.open_directory("", true, false),
+          on_click: () => navigate(""),
           active: currentPathDepth === -1,
           dndNamespace: mode,
         }),
@@ -94,7 +122,7 @@ export const PathNavigator: React.FC<Props> = React.memo(
             display: hide ? <>&bull;</> : trunc_middle(segment, 15),
             full_name: segment,
             key: i + 1,
-            on_click: (path) => actions?.open_directory(path, true, false),
+            on_click: (path) => navigate(path),
             active: is_current,
             history: is_history,
             dndNamespace: mode,
@@ -116,7 +144,7 @@ export const PathNavigator: React.FC<Props> = React.memo(
             const pathSegments = currentPath.split("/");
             pathSegments.pop();
             const parentPath = pathSegments.join("/");
-            actions?.open_directory(parentPath, true, false);
+            navigate(parentPath);
           }}
           disabled={!canGoUp}
           title={canGoUp ? "Go up one directory" : "Already at home directory"}
